packages/doc_rag_mcp/pyproject.toml
[project]
name = "doc-rag-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crawl4ai>=0.6.3",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "langchain>=0.3.25",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.22",
    "mcp[cli]>=1.9.3",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
    "requests>=2.32.4",
]
